Refactor editor structure: separate main logic and DOM elements, implement new utility functions
