+++ index.html
@@ -5,8 +5,8 @@
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Code Editor</title>
-	<link rel="stylesheet" href="src/style.css">
-	<script defer type="module" src="src/main.js"></script>
+	<link rel="stylesheet" href="src/styles/style.css">
+	<script defer type="module" src="src/editor/main.js"></script>
 </head>
 
 <body>
