+++ src/components/Caret.js
