+++ src/core/FuzzySearch.js
