+++ src/core/Lexer.js
