+++ src/core/Utility.js
