+++ src/global/Globals.js
