+++ src/secrets/VeryImportantFileDoNotTouchPleaseContainsSecretToUniverse.js
@@ -2,8 +2,8 @@
 // Since One Punch Man S3 is gonna release soon, HERES THE SURPRISEE
 // NOW HIT "CTRL + shift + o" to activate ONE PUNCH MODE!
 
-import {suggestionContainer} from "./DOMElements.js";
-import {OPMModeSettings} from "./Globals.js";
+import {suggestionContainer} from "../global/DOMElements.js";
+import {OPMModeSettings} from "../global/Globals.js";
 
 
 export function OnePunchhhhhhhhhhhhh(editor) {
